refacto actions/choix > resolveActions, actionPayante...

+++ UnePieceAngular/src/app/action/action.component.ts
@@ -13,7 +13,8 @@ import { ActionService } from '../action.service';
 export class ActionComponent {
   partie: Partie = new Partie();
   action: Action = new Action();
-  visible?: boolean = true;
+  choixPossible: boolean = true;
+  actionPayante: boolean = false;
   text?: string;
   title?: string = 'Faites Face';
   title1?: string;
@@ -42,12 +43,13 @@ export class ActionComponent {
         this.title2 = 'Fuir';
         break;
       case 'Tempete':
-        this.visible = false;
+        this.choixPossible = false;
         this.text =
           'Accrochez vous matelots, la tempête vient droit sur nous !!';
         break;
       case 'Restaurant':
         this.text = 'Qui a une petite faim ?';
+        this.actionPayante = true;
         this.title1 = 'Restaurer';
         this.title2 = 'Continuer';
         break;
@@ -69,15 +71,15 @@ export class ActionComponent {
     if (this.partie.joursRestants! > 1) {
       this.partie.joursRestants!--;
       this.partie.duree!++;
-      this.action = this.partie.actions.shift() as Action;
-     
-      this.checkForText(this.action.event?.odyssee as string);
-      this.partieService.update(this.partie).subscribe(() => {
-        this.actionService.delete(this.action.id).subscribe(() => {
-        
-          this.partieService.savePartieInStorage(this.partie);
+      if (this.partie.actions.length != 0) {
+        this.action = this.partie.actions.shift() as Action;
+        this.checkForText(this.action.event?.odyssee as string);
+        this.partieService.update(this.partie).subscribe(() => {
+          this.actionService.delete(this.action.id).subscribe(() => {
+            this.partieService.savePartieInStorage(this.partie);
+          });
         });
-      });
+      }
     } else {
       this.partie.joursRestants = this.partie.ile?.attente;
       this.partie.duree!++;
@@ -89,53 +91,126 @@ export class ActionComponent {
     this.partieService.checkEndOfGame();
   }
 
-  bouton1() {
-    console.log('(this.action.event?.odyssee as string) :>> ', (this.action.event?.odyssee as string));
-    
-    if ((this.action.event?.odyssee as string) == 'Restaurant' &&
-      this.partie.tresor! + this.action.tresor! < 0) {
-      alert("Pas assez d'argent pour cette action !");
-    } else {
-      this.partie.membres.forEach((membre, index) => {
-        membre.pv! -= this.action.degatMembre!;
-        this.partieService.checkPvMembre(membre, index);
-      });
-      this.partie.navire!.robustesse! -= this.action.degatNavire!;
-      this.partie.tresor! += this.action.tresor!;
-      if(this.partie.tresor! < 0 ) {
-        this.partie.tresor = 0;
+  resolveAction() {
+    // bouton 1
+    if (this.action.choix == true) {
+      if (
+        this.actionPayante == true &&
+        this.partie.tresor! + this.action.tresor! < 0
+      ) {
+        alert("Pas assez d'argent pour cette action !");
+      } else {
+        this.partie.membres.forEach((membre, index) => {
+          membre.pv! -= this.action.degatMembre!;
+          this.partieService.checkDeathMembre(membre, index);
+        });
+        this.partie.navire!.robustesse! -= this.action.degatNavire!;
+        this.partie.tresor! += this.action.tresor!;
+        if (this.partie.tresor! < 0) {
+          this.partie.tresor = 0;
+        }
+        this.suite();
+      }
+    }
+    // bouton 2
+    else if (this.action.choix == false) {
+      if (this.actionPayante == false) {
+        const degat: number = Math.floor(Math.random() * 4);
+        const alea: number = Math.floor(Math.random() * 3) + 1;
+        
+		// tresor
+		if (alea == 1) {
+          this.partie.tresor! -= degat * 2;
+          if (this.partie.tresor! <= 0) {
+            this.partie.tresor = 0;
+            alert('Vous avez perdu tout votre or !');
+          } else {
+            alert('Vous avez perdu ' + degat * 2 + '฿');
+          }
+        } 
+		
+		// robustesse
+		else if (alea == 2) {
+          this.partie.navire!.robustesse! -= degat;
+          alert('Votre navire prend ' + degat + ' dégâts');
+        } 
+		
+		// degat membres
+		else if (alea == 3) {
+          const membreRandom: number = Math.floor(
+            Math.random() * this.partie.membres.length
+          );
+          // 2 chance sur 3 de n'affecter qu'un membre
+          const alea2: number = Math.floor(Math.random() * 3);
+          if (alea2 != 3) {
+            this.partie.membres[membreRandom].pv! -= degat;
+            alert(
+              this.partie.membres[membreRandom].pirate?.nom +
+                ' subit ' +
+                degat +
+                ' dégâts'
+            );
+          } else {
+            this.partie.membres.forEach((membre, index) => {
+              membre.pv! -= degat;
+              this.partieService.checkDeathMembre(membre, index);
+            });
+            alert('Vos membres subissent ' + degat + ' dégâts');
+          }
+        }
       }
       this.suite();
     }
   }
 
+  bouton1() {
+    console.log(
+      '(this.action.event?.odyssee as string) :>> ',
+      this.action.event?.odyssee as string
+    );
+    this.action.choix = true;
+    this.resolveAction();
+  }
+
   bouton2() {
-    var degat: number = Math.floor(Math.random() * 4);
-    //choisir aléatoirement degatnavire, degatmembre ou tresor et enlever le degat
-    var alea: number = Math.floor(Math.random() *3)+1;
-    if(alea==1){
-      if (this.partie.tresor! == 0){
-        alert ("Tu as de la chance d'être pauvre, tu n'as plus de Berry à perdre...")
+    // choisit aléatoirement degatnavire, degatmembre ou tresor et enlève degat
+    const degat: number = Math.floor(Math.random() * 4);
+    const alea: number = Math.floor(Math.random() * 3) + 1;
+    if (alea == 1) {
+      // tresor
+      this.partie.tresor! -= degat * 2;
+      if (this.partie.tresor! <= 0) {
+        this.partie.tresor = 0;
+        alert('Vous avez perdu tout votre or !');
+      } else {
+        alert('Vous avez perdu ' + degat * 2 + '฿');
       }
-      else {
-        this.partie.tresor! -= degat*2 ;
-        alert ("Vous perdez "+(degat*2)+"฿")
-        if (this.partie.tresor! <= 0){
-          this.partie.tresor = 0;
-          alert ("Tu es maintenant pauvre, gueux.")
-        }
+    } else if (alea == 2) {
+      // robustesse
+      this.partie.navire!.robustesse! -= degat;
+      alert('Votre navire prend ' + degat + ' dégâts');
+    } else if (alea == 3) {
+      // degat membres
+      const membreRandom: number = Math.floor(
+        Math.random() * this.partie.membres.length
+      );
+      // 2 chance sur 3 de n'affecter qu'un membre
+      const alea2: number = Math.floor(Math.random() * 3);
+      if (alea2 != 3) {
+        this.partie.membres[membreRandom].pv! -= degat;
+        alert(
+          this.partie.membres[membreRandom].pirate?.nom +
+            ' subit ' +
+            degat +
+            ' dégâts'
+        );
+      } else {
+        this.partie.membres.forEach((membre, index) => {
+          membre.pv! -= degat;
+          this.partieService.checkDeathMembre(membre, index);
+        });
+        alert('Vos membres subissent ' + degat + ' dégâts');
       }
-    }
-    if(alea==2){
-      this.partie.navire!.robustesse! -= degat ;
-      alert ("Votre navire prend "+degat+" de dégâts")
-    }
-    if(alea==3){
-      this.partie.membres.forEach((membre, index) => {
-        membre.pv! -= degat;
-        this.partieService.checkPvMembre(membre, index);
-      });
-      alert ("Vos membres perdent tous "+(degat)+"pv")
     }
     this.suite();
   }
@@ -143,10 +218,10 @@ export class ActionComponent {
   boutonTempete() {
     this.partie.membres.forEach((membre, index) => {
       membre.pv! -= this.action.degatMembre!;
-      this.partieService.checkPvMembre(membre, index);
+      this.partieService.checkDeathMembre(membre, index);
     });
     this.partie.navire!.robustesse! -= this.action.degatNavire!;
-    this.visible = true;
+    this.choixPossible = true;
     this.suite();
   }
 }
